pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "matryoshka"
version = "0.1.0"
description = "Nested MGE detection and annotation for complete bacterial genomes"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.83",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "requests>=2.28",
]

[project.scripts]
matryoshka = "matryoshka.__main__:cli"
